fix: moves AccountColor to a new file "color.ts"
